pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dobot-cr-controller"
version = "0.1.0"
description = "Professional CLI tool for controlling Dobot CR series collaborative robots"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["dobot", "robotics", "cr5a", "collaborative-robot", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "pyyaml>=5.4.0",
    "numpy>=1.19.0",
    "requests>=2.25.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.scripts]
dobot-cr = "dobot_cr.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/dobot-cr-controller"
Documentation = "https://github.com/yourusername/dobot-cr-controller#readme"
Repository = "https://github.com/yourusername/dobot-cr-controller"
"Bug Tracker" = "https://github.com/yourusername/dobot-cr-controller/issues"

[tool.black]
line-length = 100
target-version = ['py37']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
